+++ tests/schema/test_response_schema.py
@@ -5,8 +5,8 @@ import pytest
 from drf_yasg import openapi
 
 import winter
-from winter.schema.generation import build_response_schema
-from winter.schema.generation import build_responses_schemas
+from winter_openapi.generation import build_response_schema
+from winter_openapi.generation import build_responses_schemas
 from winter.web.controller import get_component
 from winter.web.routing import get_route
 from ..controllers import ControllerWithExceptions